+++ packages/engine/Source/Scene/Model/MaterialPipelineStage.js
@@ -148,9 +148,21 @@ MaterialPipelineStage.process = function (
     VertexAttributeSemantic.NORMAL
   );
 
+  //Disable normals if PointCloud has them, but the parameter is set
+  //to false in PointCloudShading.normalShading.
+  let hasNormalsDisabling = false;
+  if (defined(model.pointCloudShading)) {
+    hasNormalsDisabling = !model.pointCloudShading.normalShading;
+  }
+
   // Classification models will be rendered as unlit.
   const lightingOptions = renderResources.lightingOptions;
-  if (material.unlit || !hasNormals || hasClassification) {
+  if (
+    material.unlit ||
+    !hasNormals ||
+    hasClassification ||
+    hasNormalsDisabling
+  ) {
     lightingOptions.lightingModel = LightingModel.UNLIT;
   } else {
     lightingOptions.lightingModel = LightingModel.PBR;